structure loss based on inner dist matrices
